Change Set User's Role command to both have a role to set and one to unset. Also making commands that refer to roles store the id instead of name.

+++ src/main/java/com/ryan_mtg/servobot/model/editors/CommandTableEditor.java
@@ -54,7 +54,7 @@ public class CommandTableEditor {
             Game game = Game.get((int)(long) commandRow.getLongParameter());
 
             gameQueueEditor.createGameQueue(game,
-                    savedGameQueue -> commandRow.setLongParameter(savedGameQueue.getId()));
+                    savedGameQueue -> commandRow.setLongParameter((long) savedGameQueue.getId()));
         }
 
         Command command = commandSerializer.createCommand(commandRow, bookTable.getBookMap());
